requirements.txt
numpy==1.15.0
img2pdf==0.3.1
pyocr==0.5.3
opencv_python==3.4.2.17
Pillow==5.2.0
Unidecode==1.0.22
